--- packages/atom.io/ephemeral/api-extractor.json
@@ -1,4 +1,0 @@
-{
-	"$schema": "https://developer.microsoft.com/json-schemas/api-extractor/v7/api-extractor.schema.json",
-	"extends": "../api-extractor.json"
-}
