--- packages/atom.io/ephemeral/jsdoc.json
@@ -1,21 +1,0 @@
-{
-	"tags": {
-		"allowUnknownTags": true,
-		"dictionaries": ["jsdoc"]
-	},
-	"source": {
-		"include": ["src"],
-		"includePattern": "\\.(js|ts)$",
-		"excludePattern": "(node_modules/|docs)"
-	},
-	"opts": {
-		"destination": "./docs",
-		"recurse": true,
-		"template": "templates/default",
-		"encoding": "utf8"
-	},
-	"templates": {
-		"cleverLinks": true,
-		"monospaceLinks": true
-	}
-}
