+++ packages/tsdoc.json/__tests__/builld.test.ts
@@ -1,0 +1,6 @@
+beforeEach(async () => {})
+afterEach(() => {})
+
+describe(`tsdoc.json`, () => {
+	it(`builds docs for atom.io`, async () => {})
+})
