+++ packages/tsdoc.json/package.json
@@ -1,0 +1,46 @@
+{
+	"name": "tsdoc.json",
+	"version": "0.6.2",
+	"license": "MIT",
+	"author": {
+		"name": "Jeremy Banka",
+		"email": "hello@jeremybanka.com"
+	},
+	"publishConfig": {
+		"access": "public"
+	},
+	"type": "module",
+	"files": ["bin", "dist", "src", "tsdj.schema.json"],
+	"main": "dist/tsdj.js",
+	"types": "dist/tsdj.d.ts",
+	"bin": {
+		"tsdj": "./bin/tsdj.x.js"
+	},
+	"dependencies": {
+		"@microsoft/tsdoc": "0.15.0",
+		"colors": "1.4.0"
+	},
+	"scripts": {
+		"build": "concurrently \"bun:build:*\" && bun run schema",
+		"build:lib": "rimraf dist && bun build --outdir dist --target node --external tsdj --external glob --external npmlog --external simple-git -- src/tsdj.ts ",
+		"build:bin": "rimraf bin && bun build --outdir bin --target node --external tsdj --external glob --external npmlog --external simple-git -- src/tsdj.x.ts",
+		"build:dts": "tsup",
+		"schema": "bun src/tsdj.x.ts --testPattern --testCommand --certifyCommand -- schema && mv tsdj.schema.json dist",
+		"lint:biome": "biome check -- .",
+		"lint:eslint": "eslint .",
+		"lint:types": "tsc --noEmit",
+		"lint": "bun run lint:biome && bun run lint:eslint",
+		"test": "bun test --watch -- **/*.test.ts",
+		"test:once": "bun test -- **/*.test.ts",
+		"test:coverage": "echo no test coverage yet"
+	},
+	"devDependencies": {
+		"@types/bun": "1.1.3",
+		"@types/npmlog": "7.0.0",
+		"concurrently": "8.2.2",
+		"rimraf": "5.0.7",
+		"tmp": "0.2.3",
+		"tsup": "8.1.0",
+		"typescript": "5.4.5"
+	}
+}
