+++ packages/tsdoc.json/src/index.ts
@@ -1,0 +1,3 @@
+export * from "./library"
+export * from "./namespace"
+export * from "./verbose"
