+++ packages/tsdoc.json/tsconfig.json
@@ -1,0 +1,7 @@
+{
+	"extends": "../../tsconfig.json",
+	"include": ["src", "__tests__", "**/*.config.ts"],
+	"compilerOptions": {
+		"types": ["vitest/globals", "bun"]
+	}
+}
